% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpCalomys.R
\name{CleanUpCalomys}
\alias{CleanUpCalomys}
\title{CleanUpCalomys}
\usage{
CleanUpCalomys(xls.list, id)
}
\arguments{
\item{id}{id from individuals (obtained from previous work with this DB)}

\item{raw}{output obtained from ReadCalomys, which basically just reads excel files}
}
\value{
a list with two elements, info and coord, the cleaned up version of what came in
}
\description{
This function wraps up our work on cleaning up the Calomys data.
Repeated entries are removed, all individuals have the same number of landmarks,
that kind of stuff.
This is the second step towards a nice crispy Calomys DB.
}
\details{
this function already glues left and right hemiskulls
}
\examples{
\dontrun{
}
}
\seealso{
ReadCalomys
}
\author{
Guilherme Garcia
}
