% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpCatarrhini.R
\name{CleanUpCatarrhini}
\alias{CleanUpCatarrhini}
\title{CleanUpCatarrhini}
\usage{
CleanUpCatarrhini(owm.raw, or.list)
}
\arguments{
\item{or.list}{consolidated DB from Oliveira et al (2009)}

\item{raw.list}{output from previous iteration of reading this DB}
}
\value{
a list with two elements, info and coord
}
\description{
This function cleans up Catarrhini data from excel files.
}
\details{
will GlueSkulls
}
\seealso{
ReadCatarrhine
}
\author{
Guilherme Garcia
}
