% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpHomo.R
\name{CleanUpHomo}
\alias{CleanUpHomo}
\title{CleanUpHomo}
\usage{
CleanUpHomo(xls.list)
}
\arguments{
\item{xls.list}{output obtained from ReadHomo}
}
\value{
a list with two elements, info and coord
}
\description{
This function cleans up Homo data from excel files.
}
\details{
will GlueSkulls
}
\seealso{
ReadHomo
}
\author{
Guilherme Garcia
}
