% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpPlatyrrhini.R
\name{CleanUpPlatyrrhini}
\alias{CleanUpPlatyrrhini}
\title{CleanUpPlatyrrhini}
\usage{
CleanUpPlatyrrhini(raw.list)
}
\arguments{
\item{raw.list}{raw output from reading ascii files}
}
\value{
a list with two elements, info and coord
}
\description{
This function cleans up Platyrrhini data from excel files.
}
\details{
will GlueSkulls
}
\seealso{
ReadPlatyrrhini
}
\author{
Guilherme Garcia
}
