% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpProsimian.R
\name{CleanUpProsimian}
\alias{CleanUpProsimian}
\title{CleanUpProsimian}
\usage{
CleanUpProsimian(xls.list)
}
\arguments{
\item{xls.list}{output obtained from ReadProsimian, which basically just reads
the Excel files}
}
\value{
a list with two elements, id and coord, the cleaned up version of what came in
}
\description{
This function wraps up our work on cleaning up the Prosimian data.
Repeated entries are removed, all individuals have the same number of landmarks,
that kind of stuff.
This is the second step towards a nice crispy Prosimian DB.
}
\examples{
\dontrun{
prosimian.cleanup <- CleanUpProsimian(prosimian.raw)
}
}
\seealso{
ReadProsimian
}
\author{
Guilherme Garcia, Anna Penna
}
