% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenealogyCalomys.R
\name{GenealogyCalomys}
\alias{GenealogyCalomys}
\title{GenealogyCalomys}
\usage{
GenealogyCalomys(filled)
}
\arguments{
\item{filled}{output from MissLMCalomys}
}
\value{
List with three elements:
coordinate data (coord), specimen information (info) and genealogy information (gen),
as produced by inverseA from MCMCglmm; thus, gen is itselg a list with three elements.
}
\description{
Creates inverse genealogy matrix for Calomys data.
}
\seealso{
inverseA
}
\author{
Guilherme Garcia
}
