% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlueSkull.R
\name{GlueSkull}
\alias{GlueSkull}
\title{GlueSkull}
\usage{
GlueSkull(Ai, Zi)
}
\arguments{
\item{Ai}{first set of landmark configurations}

\item{Zi}{second set of landmark configurations}
}
\value{
matrix with joined configurations
}
\description{
This function joins two hemiskulls
}
\examples{

}
\author{
Guilherme Garcia
}
