% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlueSkulls.R
\name{GlueSkulls}
\alias{GlueSkulls}
\title{GlueSkulls}
\usage{
GlueSkulls(A, Z)
}
\arguments{
\item{A}{first set of landmark configurations}

\item{Z}{second set of landmark configurations}
}
\value{
array with joined configurations
}
\description{
This function joins two different sets of landmark configuration which correspond
to the same individuals
}
\examples{

}
\author{
Guilherme Garcia
}
