% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCalomys.R
\name{ReadCalomys}
\alias{ReadCalomys}
\title{ReadCalomys}
\usage{
ReadCalomys(filename)
}
\arguments{
\item{filename}{excel register file (xls)}
}
\value{
list with two elements:
D: right side
E: left side
}
\description{
This function reads Excel files containing the Calomys colony data.
}
\examples{
\dontrun{
calomys.list <-
    dir(path = '../Raw Data/Calomys', pattern = 'input',
        recursive = TRUE, include.dirs = TRUE, full.names = TRUE)

calomys.raw <-
    alply(calomys.list, 1, function(f)
        {
            print(f)
            ReadCalomys(f)
        })

calomys.list <-
    calomys.list [!is.na(calomys.raw)]

calomys.raw <-
    calomys.raw [!is.na(calomys.raw)]

calomys.list <- gsub('../Raw Data/Calomys/', '', calomys.list)
}
}
\author{
Guilherme Garcia
}
