% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCatarrhine.R
\name{ReadCatarrhine}
\alias{ReadCatarrhine}
\title{ReadCatarrhine}
\usage{
ReadCatarrhine(filename)
}
\arguments{
\item{filename}{catarrhine register file (xls format)}
}
\value{
list with three elements:
id: data.frame with specimen information
A: array with A view coordinates
Z: array with Z view coordinates
}
\description{
This function reads Excel files with catarrhine coordinate data.
}
\examples{
\dontrun{
catarrhine.list <- 
    dir(path = '../Raw Data/Catarrhini',
        pattern = '.xls', recursive = TRUE, include.dirs = TRUE, full.names = TRUE)

catarrhine.raw <-
    alply(catarrhine.list, 1, function(f)
        {
            print(f)
            ReadCatarrhine(f)
        })

catarrhine.list <-
    catarrhine.list [!is.na(catarrhine.raw)]

catarrhine.raw <-
    catarrhine.raw [!is.na(catarrhine.raw)]

catarrhine.list <- gsub('../Raw Data/Catarrhini/', '', catarrhine.list)

names(catarrhine.raw) <- catarrhine.list

save(catarrhine.raw, file = 'Catarrhini/01_from_files.RData')
} 
}
\author{
Guilherme Garcia
}
