% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadHomo.R
\name{ReadHomo}
\alias{ReadHomo}
\title{ReadHomo}
\usage{
ReadHomo(filename)
}
\arguments{
\item{filename}{excel filename}
}
\value{
list with three elements:
id: data.frame with specimen information
A: array with A view coordinates
Z: array with Z view coordinates
}
\description{
This function reads Excel files with Homo sapiens data.
}
\examples{
\dontrun{
homo.list <- 
    dir(path = '../Raw Data/Homo/dataset- homo sapiens',
        pattern = '.xls', recursive = TRUE, include.dirs = TRUE, full.names = TRUE)

homo.raw <-
    alply(homo.list, 1, function(f)
        {
            print(f)
            ReadHomo(f)
        })

homo.list <-
    homo.list [!is.na(homo.raw)]

homo.raw <-
    homo.raw [!is.na(homo.raw)]

homo.list <- gsub('../Raw Data/Homo/dataset- homo sapiens/', '', homo.list)

names(homo.raw) <- homo.list

save(homo.raw, file = 'Homo/01_from_files.RData')
}
}
\author{
Guilherme Garcia
}
