% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadPlatyrrhini.R
\name{ReadPlatyrrhini}
\alias{ReadPlatyrrhini}
\title{ReadPlatyrrhini}
\usage{
ReadPlatyrrhini(path)
}
\arguments{
\item{path}{path to Platyrrhini raw data}
}
\value{
list with three elements
id: information
A: coordinate data for frontal view (most landmarks)
Z: coordinate data for back view (some neuro landmarks)
}
\description{
This functions read Platyrrhini POLHEMUS data, stored in ascii files
}
\details{
Although configurations are ordered, some individuals have been
measured more than once. Thus, elements in both A and Z exceed entries in id.
Individual configurations are labeled in both A and Z so that
replicates are identified, in the form (number in id).(replicate).
Also notice that the number of replicates for each A and Z may be different.
}
\examples{

ReadPlatyrrhini(path = '../Raw Data/Platyrrhini')
}
\author{
Guilherme Garcia
}
