% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinPlateSpline.R
\name{ThinPlateSpline}
\alias{ThinPlateSpline}
\title{ThinPlateSpline}
\usage{
ThinPlateSpline(target.shape, reference.shape)
}
\arguments{
\item{target.shape}{target shape}

\item{reference.shape}{reference shape}
}
\value{
a list with some elements
}
\description{
Estimates thin plate spline interpolation between reference and target shape
}
